% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reply_tweet.R
\name{reply_tweet}
\alias{reply_tweet}
\title{Replies the tweet with the new image.}
\usage{
reply_tweet(tweet)
}
\arguments{
\item{tweet}{A list with the tweet data.}
}
\description{
Replies the tweet with the ggplotized image.
}
