% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_ggplotme.R
\name{start_ggplotme}
\alias{start_ggplotme}
\title{Start ggplotme bot server.}
\usage{
start_ggplotme(
  user = "@ggplotme",
  from_time = Sys.time(),
  sleep_secs = 15 * 60
)
}
\arguments{
\item{user}{A character with the bot username.}

\item{from_time}{A POSIXct indicating minimum tweets time to attend.}

\item{sleep_secs}{Numeric indicating the seconds to sleep between execution.}
}
\description{
Starts the server, will be executing until it is aborted Ctrl+c .
}
