% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ggplot.R
\name{to_ggplot}
\alias{to_ggplot}
\title{Returns a ggplotized image.}
\usage{
to_ggplot(
  img_url,
  x_points = NA,
  y_points = NA,
  n_colors = NA,
  title = "ggplotized by @ggplotme"
)
}
\arguments{
\item{img_url}{A character with the path of the image to convert.}

\item{x_points}{A numeric indicating the new image width.}

\item{y_points}{A numeric indicating the new image height.}

\item{n_colors}{A numeric indicating the number of colors to use in the new
image.}

\item{title}{A character with the title to use in the new image.}
}
\description{
Returns the png saved file of the generated ggplot.
}
